+++ Changelog.md
@@ -2,6 +2,9 @@
 
 ## v1.25.0 
 
+**New features:**
+- Added new constants to `basicMath.h`, falling back to `math.h` constants when available
+
 **Improvements:**
 - Further improved compliance to MISRA guidelines
 - Added cppcheck configuration file